+++ src/modules/producers/services/producers.service.ts
@@ -7,6 +7,7 @@ import { CreateProducerDto } from '../dto/create-producer.dto';
 import { UpdateProducerDto } from '../dto/update-producer.dto';
 import { Producer } from '../entities/producer.entity';
 import { AlreadyExists } from 'src/shared/errors/already-exists';
+import { NotFound } from 'src/shared/errors/not-found';
 
 @Injectable()
 export class ProducersService {
@@ -39,19 +40,39 @@ export class ProducersService {
       where: { id },
     });
 
+    if (!producer) {
+      throw new NotFound('Producer not found');
+    }
+
     return plainToClass(Producer, producer);
   }
 
   async update(id: string, { name }: UpdateProducerDto) {
-    const producer = await this.prisma.producer.update({
+    const producer = await this.prisma.producer.findUnique({
+      where: { id },
+    });
+
+    if (!producer) {
+      throw new NotFound('Producer not found');
+    }
+
+    const updatedProducer = await this.prisma.producer.update({
       where: { id },
       data: { name },
     });
 
-    return plainToClass(Producer, producer);
+    return plainToClass(Producer, updatedProducer);
   }
 
   async remove(id: string) {
+    const producer = await this.prisma.producer.findUnique({
+      where: { id },
+    });
+
+    if (!producer) {
+      throw new NotFound('Producer not found');
+    }
+
     await this.prisma.producer.delete({
       where: { id },
     });
